qemu: fix pr-helper0 remain
If libvirt receive DISCONNECTED event and set prDaemonRunning to false,
and qemuDomainRemoveDiskDevice is performing in the meantime.
qemuDomainRemoveDiskDevice will return directly by prDaemonRunning
check, so the pr-helper0 object will remain.
In that case we should try harder and also kill the pr-helper process,
so qemuProcessKillManagedPRDaemon could use the same treatment.

Signed-off-by: Jie Wang <wangjie88@huawei.com>
Message-Id: <1559288011-60350-1-git-send-email-wangjie88@huawei.com>