virDomainObjListAdd: Leave def assigning as an exercise for caller
Some driver have concept of jobs (e.g. qemu, lxc, libxl and vz).
This means that whenever something wants to modify a domain
object it needs to acquire job (to mutually exclude with other
threads trying to modify the same object). And aforementioned
drivers do that more or less successfully. Except for
virDomainObjListAdd() which may modify the domain object without
acquiring job. This is harmful because if a thread that's
executing say an API in qemu driver unlock a domain object and
locks it again the object has changed even though the thread
acquired job before.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Reviewed-by: Nikolay Shirokovskiy <nshirokovskiy@virtuozzo.com>
Message-Id: <c6ccd4987599a8d8a5a863e0da0029e914e18892.1559725591.git.mprivozn@redhat.com>