qemu: Grab modify job for changing domain XML
Changing domain definition must be guarded with acquiring modify
job. The problem is if there is a thread executing say
qemuDomainSetMemoryStatsPeriod() which is an API that acquires
modify job and then possibly unlock the domain object and locks
it again. However, becasue virDomainObjAssignDef() does not take
a job (only object lock) it may have changed the domain
definition while the other thread unlocked the domain object in
order to talk on the monitor. For instance:

Thread1:
1) lookup domain object and lock it
2) acquire job
3) get pointers to live and persistent defs
4) unlock the domain object
5) talk to qemu on the monitor

Thread2 (Execute qemuDomainDefineXMLFlags):
1) lookup domain object and lock it
2) virDomainObjAssignDef() which overwrites persistent def and
   frees the old one
3) unlock domain object

Thread1:
6) lock the domain object again
7) try to access the persistent def via pointer stored in 3)

Now, this will crash because the pointer from step 3) points to a
memory that was freed.

However, if Thread2 waited and acquired modify job (just like
Thread1) then these two would serialize and no harm would be
done.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <d408dbe8dc0e342dbefdde46e20e5faba5068c21.1559725591.git.mprivozn@redhat.com>