virsh: undefine: Rename --delete-snapshots to --delete-storage-volume-snapshots
The old flag name confused some users into thinking it's the correct way
to undefine a VM with libvirt (not storage volume) snapshots.

The correct flag in that case is way less obvious: --snapshots-metadata.

Rename the flag (by adding an alias) to something which will promote
looking up the actual purpose of the flag.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <8ebb9d64ec8fde2164356bed83fa66a93aeb5ccd.1559733766.git.pkrempa@redhat.com>
Reviewed-by: Fabiano Fidêncio <fidencio@redhat.com>