backup: Prepare for Unix sockets in QMP nbd-server-start
Migration always uses a TCP socket for NBD servers, because we don't
support same-host migration. But upcoming pull-mode incremental backup
needs to also support a Unix socket, for retrieving the backup from
the same host. Support this by plumbing virStorageNetHostDef through
the monitor calls, since that is a nice reusable struct that can track
both TCP and Unix sockets.

Update qemumonitorjsontest to not only test the response to the
command, but to actually verify that the command itself uses the two
correct QMP forms.  I'm actually a bit surprised that we are not
utilizing qemuMonitorTestAddItemVerbatim more frequently.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190606030110.4457-2-eblake@redhat.com>