tests: Normalize quotes for *-graphics
Right now *-headless and *-graphics tests are using different
quoting styles, which results in the diff between them being
basically useless, whereas we would like it to be possible to
compare these files directly and easily spot the differences.

Convert all *-graphics tests to single quotes, which is the
overwhelmingly more popular style and also the one libvirt
itself uses when formatting XML, which is a fact that will
come in handy later.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Message-Id: <20190614080439.29627-2-abologna@redhat.com>