qemu: driver: Remove support for native->XML conversion
This code is really neglected and does not at all work reliably. It
can't even be used for converting our own commandline back.

Since this was mostly useful for aiding migration from manually run qemu
to libvirt and will not work for this puspose in many cases it's not
worth having in my opinion.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <27f45340fc315f8f1a9ee42b9ad584578b1493bd.1560518789.git.pkrempa@redhat.com>
Reviewed-by: Ján Tomko <jtomko@redhat.com>