build: Require cppi for syntax-check
Since its introduction in v0.8.0~314 we allowed for cppi to be
not present for 'syntax-check' because the package did exist in
Fedora Core 12. Well, we don't care anymore ¯\_(ツ)_/¯
All recent distros have it. Moreover, it's fairly easy to miss
'cppi not installed' message in sea of syntax-check output (esp.
for newbies who probably benefit from it the most).

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <e94a740431b2615470cbed763467f2a73ef02bfe.1560604642.git.mprivozn@redhat.com>
Reviewed-by: Ján Tomko <jtomko@redhat.com>