Produce more verbose error if cppi not found
It's fairly easy (especially for new contributors) to not spot
the 'cppi not installed' line in the syntax-check output. Turn it
into a banner that is more visible and at the same time add it as
a build dependency. Unfortunately, RHEL doesn't ship cppi so we
can add the dependency only for Fedora.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <d23619de4c2645aee1bf5150e90b2a767c419a7c.1560773541.git.mprivozn@redhat.com>