Produce more verbose error if cppi not found
It's fairly easy (especially for new contributors) to not spot
the 'cppi not installed' line in the syntax-check output. Turn it
into a banner that is more visible and at the same time add it as
a build dependency. Unfortunately, RHEL doesn't ship cppi so we
can add the dependency only for Fedora.

Since it's v1 this has effectively became code copied over from
Andrea's review suggestions.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <8aeac048b83ad6de863ae87a645505b92ef6013a.1560788460.git.mprivozn@redhat.com>