test_driver: consider DHCP ranges in testDomainInterfaceAddresses
testDomainInterfaceAddresses always returns the same hard-coded
addresses. Change the behavior such as if there is a DHCP range defined,
addresses are returned from that pool.

The specific address returned depends on both the domain id and the
specific guest interface in an attempt to return unique addresses *most
of the time*.

Additionally, return IPv6 addresses too when needed.

Signed-off-by: Ilias Stamatis <stamatis.iliass@gmail.com>
Message-Id: <20190619111817.19562-4-stamatis.iliass@gmail.com>