docs: Use our own implementation for fetching the RSS data
This is just a small script I wrote.  It works the same way as all the libraries
together which we are bundling, but with just JS.  The only difference is that
the day of the date is formatted as 2-digit, but this should be a bug in
Firefox (at least locally for me) as the documentation states that type
'numeric' should actually not have leading zeros.

It will not be executed when the page is loaded locally.

It also uses the same proxy that query-rss uses.  We can get rid of that, but
we'd either need to have our own proxy, send a sss header to allow fetching the
atom.xml or providing JSONP access to the RSS feed on virt-planet.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>
Message-Id: <a5ab6539e0102aad56a8f16a57e2c5615a664c4f.1560953465.git.mkletzan@redhat.com>