Don't parse/format vram attribute for cirrus video
Since the cirrus vga memory size isn't configurable, we can ignore any
'vram' attribute when parsing a domain definition. However, when no
value is specified, it ends up getting set to a default value of 16MB.
This 16MB value is not used anywhere (for example, it is not passed as
an argument to qemu), but is displayed in the XML definition. So by
changing this default value to 0, it will also be omitted from the XML
definition of the domain.

Fixes: rhbz#1447831

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>
Message-Id: <20190620195112.26695-1-jjongsma@redhat.com>