locking, logging: put a strong dep from admin socket to main socket
It doesn't make sense to have the admin socket active if the main
socket is not running, so bind their lifecycle together.

This ensures that if primary socket is stopped, the corresponding
admin socket is also stopped.

In the reverse, starting the admin socket will also automatically
start the primary socket.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>
Message-Id: <20190627095452.28217-2-berrange@redhat.com>