util: add VIR_AUTOSTRUCT for directly calling a struct free function
The current VIR_AUTOPTR macro assumes that the struct needs to have a
auto-free function auto-generated to call the real free function.

The new VIR_AUTOSTRUCT macro allows for structs which already have a
free function which takes a pointer to a pointer to the struct and can
thus be used directly for auto-cleanup.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>
Message-Id: <20190627095452.28217-5-berrange@redhat.com>