tests: Add more DO_TEST_CAPS_*() macros
Right now we have macros such as DO_TEST_CAPS_LATEST_PARSE_ERROR()
and DO_TEST_CAPS_ARCH_VER(), but there is no concise way to say
"using this version of QEMU on this architecture will result in a
failure".

This commit adds

  DO_TEST_CAPS_ARCH_LATEST_FAILURE()
  DO_TEST_CAPS_ARCH_VER_FAILURE()
  DO_TEST_CAPS_ARCH_LATEST_PARSE_ERROR()
  DO_TEST_CAPS_ARCH_VER_PARSE_ERROR()

and also slightly reworks

  DO_TEST_CAPS_LATEST_PARSE_ERROR()
  DO_TEST_CAPS_ARCH_VER()

so that they use the closest more specific macro in their
implementation instead of going further back towards the more
generic building blocks.

The ordering is also altered so that it's both more logical and
fully consistent between xml2argv and xml2xml.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Message-Id: <20190702152343.30295-2-abologna@redhat.com>